fix: crash in editself when no user avatar is available

+++ server/api/src/schema/User/resolver.js
@@ -94,7 +94,7 @@ module.exports = {
         if (user.avatar) {
           // user already has an avatar, we need to delete it
           const oldAvatar = await context.prisma.file({ uri: user.avatar });
-          await deleteFile(oldAvatar.id, context);
+          if (oldAvatar) await deleteFile(oldAvatar.id, context);
         }
 
         // store file and create database entry
